Add suggestions from code review
